Added links to Chris Done's HIndent.
HIndent (https://github.com/chrisdone/hindent/) is a wonderful Haskell
pretty-printer. It takes the tedium out of doing indentation in Haskell. Highly
recommend. It includes a vim extension (or whatever it's called) and an Emacs
minor mode.

It is not included in any of the links in the README. I opened an issue on the
Emacs Haskell
tutorial (https://github.com/serras/emacs-haskell-tutorial/issues/20), but this
can serve as a temporary patch for the time being.