Updates the Portuguese translation of the guide with the English version
The Portuguese version of the guide was still suggesting to install
cabal and haskell platform.

This commit updates the content of the Portuguese guide to the same
content that is provided on the README.md.

One thing I've noticed with the end result is that we are now linking to
the other documents instead of having the content on the same document,
but I believe this was refactored out from the main document after the
translation.